backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
motor
python-dotenv
pydantic
pydantic-settings
PyMuPDF
python-multipart
aiofiles
pillow
pytesseract